+++ Lambdas/GHHooks/MessageLookupRepo/MessageLookupRepo.swift
@@ -1,5 +1,5 @@
 protocol MessageLookupRepo: Sendable {
-    func getMessageID(repoID: Int, number: Int) async throws -> String
-    func markAsUnavailable(repoID: Int, number: Int) async throws
-    func saveMessageID(messageID: String, repoID: Int, number: Int) async throws
+    func getMessageID(repoID: Int64, number: Int) async throws -> String
+    func markAsUnavailable(repoID: Int64, number: Int) async throws
+    func saveMessageID(messageID: String, repoID: Int64, number: Int) async throws
 }
